Added in OGL Renderer and Config parsing

+++ Renderer/Renderer.h
@@ -4,6 +4,7 @@
 #include <GL/glew.h>
 #include <string>
 
+//Structure to make it easier to describe a renderer
 struct RendererDesc
 {
 	std::string Type;
@@ -11,14 +12,20 @@ struct RendererDesc
 	int BackBufferHeight;
 };
 
+//Pure virtual interface for a renderer. All renderers should inheirt from this
 class IRenderer
 {
 public:
 	virtual ~IRenderer() {};
 
+	//Create Renderer
 	virtual bool Create(RendererDesc &desc,SDL_Window * window) = 0;
 
-	virtual void Clear(int r, int g, int b) = 0;
+	//Clear current display buffer
+	virtual void Clear(float r, float g, float b) = 0;
+	
+	//begin rendering
 	virtual void Begin() = 0;
+	//end rendering
 	virtual void End() = 0;
 };
